# Memmap: Optimize UpdateLogicalMemory by merging mappings