Instead of creating many 128 KiB mappings, we can create a few large mappings. On my Windows PC, this speeds up GameCube (FakeVMEM) game boot times by about 200 ms and Wii game boot times by about 60 ms. Loading savestates is also faster, by about 45 ms for GameCube (FakeVMEM) games and 5 ms for Wii games. The impact is presumably smaller on other OSes because Windows is particularly slow at creating mappings.